Remove `hero_text` special case.
Previously this just added a paragraph to the top of the doc. We can get
the same exact result in Markdown, without any special syntax or
template.

Also remove the unused `subtitle` attribute and replace fancy quotes to
simple quotes in the source. (Jekyll automatically converts to fancy
quotes in the HTML.)

User-visible impact: this removes extraneous, unintended vertical space at the top of each page due to an empty `<p></p>` tag.

Signed-off-by: Mark Lodato <lodato@google.com>