Automatically render first paragraph as "subtitle"
Previously, many pages started with a `<div class="subtitle">` that
rendered the first paragraph in larger font, to give an overview of the
page. But this was not used consistently since some pages forgot the
tag.

Now the first paragraph is always rendered larger (using CSS selectors)
without need for the tag. This makes the source easier to maintain and
also makes the site more consistent.

Signed-off-by: Mark Lodato <lodato@google.com>