Add description across SLSA pages.
Link unfurling and SEO currently only uses the default description for
the entire webpage. In order to provide more clarity to this, I unified
pages' frontmatter to specify the title and provide a description.

Resolves #693

Blog posts don't need to have descriptions added as they seem to be
handled differently. Their description is populated by the initial
content in the post.

Signed-off-by: arewm <arewm@users.noreply.github.com>