Move use cases to versioned spec directory
Now the same exact file is copied to all three /spec/v*/ directories.
The only change from the original is to update the link to example.md,
which I kept at the top level. (We can always move that one later if
needed.)

This also fixes the link for the v1.0 nav bar, which is currently
broken (it already points to the path added in this change.)

Signed-off-by: Mark Lodato <lodato@google.com>