Set `noindex` on 'onepage' pages
These pages duplicate content found elsewhere on the site, which is
generally discouraged by search engines. Since they combine multiple
pages, we can't provide a single canonical link, and because these are
likely not the pages we want to show up in search results, just mark
these as `noindex` instead.