Explicitly mention identifying package by digest
In SLSA v0.1 we had said that the provenance MUST identify the package
by cryptographic digest, but this was implicit in v1.0 from
"unambiguously identifies the output package" and "equivalent
information as SLSA Provenance". Still, there is room for
misinterpretation, so this change calls out identifying the output
artifact by cryptographic digest so there is no room for
misinterpretation.

Also splits the paragraphs to be more readable.

Fixes #601.

Signed-off-by: Mark Lodato <lodato@google.com>