Avoid MUST in "verifying artifacts" and "distributing provenance"
As discussed at the 2023-04-03 Specification Meeting, we agreed to
downgrade the requirements (MUST) in "verifying artifacts" and
"distributing provenance" to be recommendations (SHOULD) since we have
not yet sufficiently defined what a "SLSA-conformant system" means.
Hopefully we can address this in a future revision of the specification.

This leaves one use of "MUST", namely "Producers MUST publish
attestations in at least one place", because that is a requirement from
"producing artifacts".

Fixes #786.

Signed-off-by: Mark Lodato <lodato@google.com>